requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altgraph==0.17.2
antiorm==1.2.1
anyio==3.6.2
appdirs==1.4.4
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.1.0
binaryornot==0.4.4
certifi==2022.9.24
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.4
cookiecutter==2.6.0
customtkinter==5.2.2
darkdetect==0.8.0
db==0.1.1
db-sqlite3==0.0.1
decorator==5.1.1
distlib==0.3.1
docutils==0.19
et-xmlfile==1.1.0
filelock==3.0.12
Flask==2.1.2
flet==0.27.5
flet-cli==0.27.5
flet-core==0.5.2
flet-desktop==0.27.5
frozenlist==1.4.0
future==0.18.2
geocoder==1.38.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Kivy==2.1.0
kivy-deps.angle==0.3.2
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
kivymd==1.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
mccabe==0.6.1
mdurl==0.1.2
multidict==6.0.4
numpy==1.23.3
oauthlib==3.2.2
openai==0.28.1
openpyxl==3.0.10
packaging==23.1
pandas==1.5.0
peewee==3.14.10
pefile==2021.9.3
pillow==10.4.0
platformdirs==2.5.2
Pygments==2.19.1
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.4
pyperclip==1.9.0
pypiwin32==223
pypng==0.20220715.0
PySimpleGUI==4.59.0
python-dateutil==2.8.2
python-slugify==8.0.4
python-tkdnd==0.2.1
pytube==12.1.0
pytz==2021.1
pywin32==304
pywin32-ctypes==0.2.0
PyYAML==6.0.2
qrcode==7.4.2
ratelim==0.1.6
repath==0.9.0
requests==2.28.1
rfc3986==1.5.0
rich==14.0.0
six==1.15.0
sniffio==1.3.0
sqlite-web==0.4.0
sqlparse==0.4.1
text-unidecode==1.3
tkintermapview==1.29
toml==0.10.2
tqdm==4.66.1
ttkwidgets==0.13.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2022.1
urllib3==1.26.12
virtualenv==20.14.1
virtualenvwrapper-win==1.2.7
watchdog==4.0.2
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.1.2
win10toast==0.9
wrapt==1.12.1
yarl==1.9.2
